html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-window-open.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: location.replace</title>
<iframe name="iframe" data-src="support/blank.htm?nav" src="support/blank.htm?src" loading="lazy"></iframe>
<script>
const iframe = document.querySelector('iframe');
window.open(iframe.dataset.src, 'iframe');
</script>
<!-- Loading testharness.js here is intentional. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
  done();
}
</script>
